update with sources from 0.77 download
