+++ include/bliss/digraph.hh
@@ -24,11 +24,11 @@
 namespace bliss {
 
 /**
- * \brief The class for undirected, vertex colored graphs.
+ * \brief The class for directed, vertex colored graphs.
  *
- * Multiple edges between vertices are not allowed (i.e., are ignored).
+ * Multiple edges between vertices are not allowed (copies will be ignored).
  */
-class Graph : public AbstractGraph
+class Digraph : public AbstractGraph
 {
 public:
   /**
@@ -69,16 +69,17 @@ protected:
   public:
     Vertex();
     ~Vertex();
-    void add_edge(const unsigned int other_vertex);
+    void add_edge_to(const unsigned int dest_vertex);
+    void add_edge_from(const unsigned int source_vertex);
     void remove_duplicate_edges(std::vector<bool>& tmp);
     void sort_edges();
-
     unsigned int color;
-    std::vector<unsigned int> edges;
-    unsigned int nof_edges() const {return edges.size(); }
+    std::vector<unsigned int> edges_out;
+    std::vector<unsigned int> edges_in;
+    unsigned int nof_edges_in() const {return edges_in.size(); }
+    unsigned int nof_edges_out() const {return edges_out.size(); }
   };
   std::vector<Vertex> vertices;
-  void sort_edges();
   void remove_duplicate_edges();
 
   /** \internal
@@ -86,42 +87,52 @@ protected:
    * Returns the color of the vertex.
    * Time complexity: O(1).
    */
-  static unsigned int vertex_color_invariant(const Graph* const g,
+  static unsigned int vertex_color_invariant(const Digraph* const g,
                                              const unsigned int v);
-
   /** \internal
    * Partition independent invariant.
-   * Returns the degree of the vertex.
+   * Returns the indegree of the vertex.
    * DUPLICATE EDGES MUST HAVE BEEN REMOVED BEFORE.
    * Time complexity: O(1).
    */
-  static unsigned int degree_invariant(const Graph* const g,
-                                       const unsigned int v);
-
+  static unsigned int indegree_invariant(const Digraph* const g,
+                                         const unsigned int v);
+  /** \internal
+   * Partition independent invariant.
+   * Returns the outdegree of the vertex.
+   * DUPLICATE EDGES MUST HAVE BEEN REMOVED BEFORE.
+   * Time complexity: O(1).
+   */
+  static unsigned int outdegree_invariant(const Digraph* const g,
+                                          const unsigned int v);
   /** \internal
    * Partition independent invariant.
    * Returns 1 if there is an edge from the vertex to itself, 0 if not.
    * Time complexity: O(k), where k is the number of edges leaving the vertex.
    */
-  static unsigned int selfloop_invariant(const Graph* const g,
+  static unsigned int selfloop_invariant(const Digraph* const g,
                                          const unsigned int v);
 
-
-  bool refine_according_to_invariant(unsigned int (*inv)(const Graph* const g,
+  /** \internal
+   * Refine the partition \a p according to
+   * the partition independent invariant \a inv.
+   */
+  bool refine_according_to_invariant(unsigned int (*inv)(const Digraph* const g,
                                                          const unsigned int v));
 
   /*
    * Routines needed when refining the partition p into equitable
    */
-  bool split_neighbourhood_of_unit_cell(Partition::Cell * const);
-  bool split_neighbourhood_of_cell(Partition::Cell * const);
+  bool split_neighbourhood_of_unit_cell(Partition::Cell* const);
+  bool split_neighbourhood_of_cell(Partition::Cell* const);
+
 
   /** \internal
    * \copydoc AbstractGraph::is_equitable() const
    */
   bool is_equitable() const;
 
-  /* Splitting heuristics, documented in more detail in graph.cc */
+  /* Splitting heuristics, documented in more detail in the cc-file. */
   SplittingHeuristic sh;
   Partition::Cell* find_next_cell_to_be_splitted(Partition::Cell *cell);
   Partition::Cell* sh_first();
@@ -130,7 +141,6 @@ protected:
   Partition::Cell* sh_first_max_neighbours();
   Partition::Cell* sh_first_smallest_max_neighbours();
   Partition::Cell* sh_first_largest_max_neighbours();
-
 
   /* A data structure used in many functions.
    * Allocated only once to reduce allocation overhead,
@@ -142,7 +152,7 @@ protected:
 
   void initialize_certificate();
 
-
+  void sort_edges();
 
   bool nucr_find_first_component(const unsigned int level);
   bool nucr_find_first_component(const unsigned int level,
@@ -150,19 +160,16 @@ protected:
                                  unsigned int& component_elements,
                                  Partition::Cell*& sh_return);
 
-
-
-
 public:
   /**
-   * Create a new graph with \a N vertices and no edges.
+   * Create a new directed graph with \a N vertices and no edges.
    */
-  Graph(const unsigned int N = 0);
+  Digraph(const unsigned int N = 0);
 
   /**
    * Destroy the graph.
    */
-  ~Graph();
+  ~Digraph();
 
   /**
    * Read the graph from the file \a fp in a variant of the DIMACS format.
@@ -171,31 +178,29 @@ public:
    * Note that in the DIMACS file the vertices are numbered from 1 to N while
    * in this C++ API they are from 0 to N-1.
    * Thus the vertex n in the file corresponds to the vertex n-1 in the API.
-   *
    * \param fp      the file stream for the graph file
    * \param errstr  if non-null, the possible error messages are printed
    *                in this file stream
-   * \return        a new Graph object or 0 if reading failed for some
+   * \return        a new Digraph object or 0 if reading failed for some
    *                reason
    */
-  static Graph* read_dimacs(FILE* const fp, FILE* const errstr = stderr);
+  static Digraph* read_dimacs(FILE* const fp, FILE* const errstr = stderr);
 
   /**
-   * Write the graph to a file in a variant of the DIMACS format.
-   * See the <a href="https://users.aalto.fi/tjunttil/bliss">bliss website</a>
-   * for the definition of the file format.
+   * \copydoc AbstractGraph::write_dimacs(FILE * const fp)
    */
   void write_dimacs(FILE* const fp);
 
+
   /**
-   * \copydoc AbstractGraph::write_dot(FILE * const fp)
+   * \copydoc AbstractGraph::write_dot(FILE *fp)
    */
-  void write_dot(FILE* const fp);
+  void write_dot(FILE * const fp);
 
   /**
    * \copydoc AbstractGraph::write_dot(const char * const file_name)
    */
-  void write_dot(const char* const file_name);
+  void write_dot(const char * const file_name);
 
 
 
@@ -210,36 +215,17 @@ public:
   unsigned int get_nof_vertices() const {return vertices.size(); }
 
   /**
-   * \copydoc AbstractGraph::permute(const unsigned int* const perm) const
-   */
-  Graph* permute(const unsigned int* const perm) const;
-  /**
-   * \copydoc AbstractGraph::permute(const std::vector<unsigned int>& perm) const
-   */
-  Graph* permute(const std::vector<unsigned int>& perm) const;
-
-  /**
-   * \copydoc AbstractGraph::is_automorphism(unsigned int* const perm) const
-   */
-  bool is_automorphism(unsigned int* const perm) const;
-
-  /**
-   * \copydoc AbstractGraph::is_automorphism(const std::vector<unsigned int>& perm) const
-   */
-  bool is_automorphism(const std::vector<unsigned int>& perm) const;
-
-  /**
-   * Add a new vertex with color \a color in the graph and return its index.
+   * Add a new vertex with color 'color' in the graph and return its index.
    */
   unsigned int add_vertex(const unsigned int color = 0);
 
   /**
-   * Add an edge between vertices \a v1 and \a v2.
-   * Duplicate edges between vertices are ignored but try to avoid introducing
+   * Add an edge from the vertex \a source to the vertex \a target.
+   * Duplicate edges are ignored but try to avoid introducing
    * them in the first place as they are not ignored immediately but will
    * consume memory and computation resources for a while.
    */
-  void add_edge(const unsigned int v1, const unsigned int v2);
+  void add_edge(const unsigned int source, const unsigned int target);
 
   /**
    * \copydoc AbstractGraph::get_color(const unsigned int vertex) const
@@ -247,14 +233,14 @@ public:
   unsigned int get_color(const unsigned int vertex) const;
 
   /**
-   * Change the color of the vertex \a vertex to \a color.
+   * Change the color of the vertex 'vertex' to 'color'.
    */
   void change_color(const unsigned int vertex, const unsigned int color);
 
   /**
    * Get a copy of the graph.
    */
-  Graph* copy() const;
+  Digraph* copy() const;
 
   /**
    * Compare this graph to the \a other graph in a total orger on graphs.
@@ -262,7 +248,7 @@ public:
    *         -1 if this graph is "smaller than" the other, and
    *         1 if this graph is "greater than" the other.
    */
-  int cmp(Graph& other);
+  int cmp(Digraph& other);
 
   /**
    * Set the splitting heuristic used by the automorphism and canonical
@@ -273,9 +259,27 @@ public:
    * canonical versions, be sure to use the same splitting heuristics
    * for both graphs.
    */
-  void set_splitting_heuristic(const SplittingHeuristic shs) {sh = shs; }
+  void set_splitting_heuristic(SplittingHeuristic shs) {sh = shs; }
 
+  /**
+   * \copydoc AbstractGraph::permute(const unsigned int* const perm) const
+   */
+  Digraph* permute(const unsigned int* const perm) const;
 
+  /**
+   * \copydoc AbstractGraph::permute(const std::vector<unsigned int>& perm) const
+   */
+  Digraph* permute(const std::vector<unsigned int>& perm) const;
+
+  /**
+   * \copydoc AbstractGraph::is_automorphism(unsigned int* const perm) const
+   */
+  bool is_automorphism(unsigned int* const perm) const;
+
+  /**
+   * \copydoc AbstractGraph::is_automorphism(const std::vector<unsigned int>& perm) const
+   */
+  bool is_automorphism(const std::vector<unsigned int>& perm) const;
 };
 
 } // namespace bliss
