+++ include/bliss/uintseqhash.hh
@@ -1,14 +1,11 @@
-#ifndef BLISS_UINTSEQHASH_HH
-#define BLISS_UINTSEQHASH_HH
-
-#include <cstdio>
+#pragma once
 
 /*
-  Copyright (c) 2003-2015 Tommi Junttila
+  Copyright (c) 2003-2021 Tommi Junttila
   Released under the GNU Lesser General Public License version 3.
-  
+
   This file is part of bliss.
-  
+
   bliss is free software: you can redistribute it and/or modify
   it under the terms of the GNU Lesser General Public License as published by
   the Free Software Foundation, version 3 of the License.
@@ -24,8 +21,8 @@
 
 namespace bliss {
 
-/** \internal
- * \brief A hash for sequences of unsigned ints.
+/**
+ * \brief A updatable hash for sequences of unsigned ints.
  */
 class UintSeqHash
 {
@@ -35,7 +32,7 @@ public:
   UintSeqHash() {h = 0; }
   UintSeqHash(const UintSeqHash &other) {h = other.h; }
   UintSeqHash& operator=(const UintSeqHash &other) {h = other.h; return *this; }
-  
+
   /** Reset the hash value. */
   void reset() {h = 0; }
 
@@ -52,14 +49,12 @@ public:
     return (h < other.h)?-1:((h == other.h)?0:1);
   }
   /** An abbreviation for cmp(other) < 0 */
-  bool is_lt(const UintSeqHash &other) const {return(cmp(other) < 0); }
+  bool is_lt(const UintSeqHash &other) const {return cmp(other) < 0; }
   /** An abbreviation for cmp(other) <= 0 */
-  bool is_le(const UintSeqHash &other) const {return(cmp(other) <= 0); }
+  bool is_le(const UintSeqHash &other) const {return cmp(other) <= 0; }
   /** An abbreviation for cmp(other) == 0 */
-  bool is_equal(const UintSeqHash &other) const {return(cmp(other) == 0); }
+  bool is_equal(const UintSeqHash &other) const {return cmp(other) == 0; }
 };
 
 
 } // namespace bliss
-
-#endif
