+++ src/defs.cc
@@ -1,13 +1,9 @@
-#include <cstdlib>
-#include <cstdio>
-#include "bliss/defs.hh"
-
 /*
-  Copyright (c) 2003-2015 Tommi Junttila
+  Copyright (c) 2003-2021 Tommi Junttila
   Released under the GNU Lesser General Public License version 3.
-  
+
   This file is part of bliss.
-  
+
   bliss is free software: you can redistribute it and/or modify
   it under the terms of the GNU Lesser General Public License as published by
   the Free Software Foundation, version 3 of the License.
@@ -21,18 +17,10 @@
   along with bliss.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-namespace bliss {
+#include <cstdlib>
+#include <cstdio>
+#include "bliss/defs.hh"
 
-void
-fatal_error(const char* fmt, ...)
-{
-  va_list ap;
-  va_start(ap, fmt);
-  fprintf(stderr,"Bliss fatal error: ");
-  vfprintf(stderr, fmt, ap);
-  fprintf(stderr, "\nAborting!\n");
-  va_end(ap);
-  exit(1);
-}
+namespace bliss {
 
 }
