+++ src/utils.cc
@@ -1,13 +1,9 @@
-#include <cassert>
-#include <vector>
-#include "bliss/utils.hh"
-
 /*
-  Copyright (c) 2003-2015 Tommi Junttila
+  Copyright (c) 2003-2021 Tommi Junttila
   Released under the GNU Lesser General Public License version 3.
-  
+
   This file is part of bliss.
-  
+
   bliss is free software: you can redistribute it and/or modify
   it under the terms of the GNU Lesser General Public License as published by
   the Free Software Foundation, version 3 of the License.
@@ -21,72 +17,60 @@
   along with bliss.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+#include <cassert>
+#include <vector>
+#include "bliss/utils.hh"
+
 namespace bliss {
 
-void
+size_t
 print_permutation(FILE* const fp,
-		  const unsigned int N,
-		  const unsigned int* perm,
-		  const unsigned int offset)
+    const unsigned int N,
+    const unsigned int* perm,
+    const unsigned int offset)
 {
   assert(N > 0);
   assert(perm);
-  for(unsigned int i = 0; i < N; i++) {
-    unsigned int j = perm[i];
-    if(j == i)
-      continue;
-    bool is_first = true;
-    while(j != i) {
-      if(j < i) {
-        is_first = false;
-        break;
-      }
-      j = perm[j];
+  size_t r = 0;
+  unsigned int nof_cycles = 0;
+  std::vector<bool> seen(N, false);
+  for(unsigned int first = 0; first < N; first++) {
+    if(seen[first] or perm[first] == first) continue;
+    nof_cycles++;
+    r += fprintf(fp, "(%u", first+offset);
+    for(unsigned int i = perm[first]; i != first; i = perm[i]) {
+      seen[i] = true;
+      r += fprintf(fp, ",%u", i+offset);
     }
-    if(!is_first)
-      continue;
-    fprintf(fp, "(%u,", i+offset);
-    j = perm[i];
-    while(j != i) {
-      fprintf(fp, "%u", j+offset);
-      j = perm[j];
-      if(j != i)
-        fprintf(fp, ",");
-    }
-    fprintf(fp, ")");
+    r += fprintf(fp, ")");
   }
+  if(nof_cycles == 0)
+    r += fprintf(fp, "()");
+  return r;
 }
 
-void
+size_t
 print_permutation(FILE* const fp,
-		  const std::vector<unsigned int>& perm,
-		  const unsigned int offset)
+    const std::vector<unsigned int>& perm,
+    const unsigned int offset)
 {
   const unsigned int N = perm.size();
-  for(unsigned int i = 0; i < N; i++) {
-    unsigned int j = perm[i];
-    if(j == i)
-      continue;
-    bool is_first = true;
-    while(j != i) {
-      if(j < i) {
-        is_first = false;
-        break;
-      }
-      j = perm[j];
+  size_t r = 0;
+  unsigned int nof_cycles = 0;
+  std::vector<bool> seen(N, false);
+  for(unsigned int first = 0; first < N; first++) {
+    if(seen[first] or perm[first] == first) continue;
+    nof_cycles++;
+    r += fprintf(fp, "(%u", first+offset);
+    for(unsigned int i = perm[first]; i != first; i = perm[i]) {
+      seen[i] = true;
+      r += fprintf(fp, ",%u", i+offset);
     }
-    if(!is_first)
-      continue;
-    fprintf(fp, "(%u,", i+offset);
-    j = perm[i];
-    while(j != i) {
-      fprintf(fp, "%u", j+offset);
-      j = perm[j];
-      if(j != i)
-        fprintf(fp, ",");
-    }
-    fprintf(fp, ")");
+    r += fprintf(fp, ")");
   }
+  if(nof_cycles == 0)
+    r += fprintf(fp, "()");
+  return r;
 }
 
 bool
